+++ e2e/nx-boot-gradle-e2e/tests/nx-boot-gradle.spec.ts
@@ -124,7 +124,7 @@ describe('nx-boot-gradle e2e', () => {
 
     expect(() =>
       checkFilesExist(
-        `node_modules/@jnxplus/tools/linters/checkstyle/checkstyle-${checkstyleVersion}.jar`,
+        `node_modules/@jnxplus/tools/linters/checkstyle/checkstyle-${checkstyleVersion}-all.jar`,
         `node_modules/@jnxplus/tools/linters/ktlint/ktlint`
       )
     ).not.toThrow();