+++ e2e/nx-quarkus-gradle-e2e/tests/nx-quarkus-gradle.spec.ts
@@ -126,7 +126,7 @@ describe('nx-quarkus-gradle e2e', () => {
 
     expect(() =>
       checkFilesExist(
-        `node_modules/@jnxplus/tools/linters/checkstyle/checkstyle-${checkstyleVersion}.jar`,
+        `node_modules/@jnxplus/tools/linters/checkstyle/checkstyle-${checkstyleVersion}-all.jar`,
         `node_modules/@jnxplus/tools/linters/ktlint/ktlint`
       )
     ).not.toThrow();